+++ src/main/java/woowacourse/shoppingcart/domain/customer/Nickname.java
@@ -1,23 +1,23 @@
-package woowacourse.shoppingcart.domain;
+package woowacourse.shoppingcart.domain.customer;
 
 import java.util.Objects;
 import java.util.regex.Pattern;
-import woowacourse.shoppingcart.exception.IllegalEmailException;
+import woowacourse.shoppingcart.exception.bodyexception.IllegalNicknameException;
 
-public class Email {
+public class Nickname {
 
-    private static final String REGEX = "^[a-zA-Z0-9_!#$%&'*+/=?`{|}~^.-]+@[a-zA-Z0-9.-]+$";
+    private static final String REGEX = "[a-zA-Z0-9가-힣]{2,8}";
 
     private final String value;
 
-    public Email(String value) {
+    public Nickname(String value) {
         validate(value);
         this.value = value;
     }
 
     private void validate(String value) {
         if (!Pattern.matches(REGEX, value)) {
-            throw new IllegalEmailException();
+            throw new IllegalNicknameException();
         }
     }
 
@@ -33,8 +33,8 @@ public class Email {
         if (o == null || getClass() != o.getClass()) {
             return false;
         }
-        Email email = (Email) o;
-        return Objects.equals(value, email.value);
+        Nickname nickname = (Nickname) o;
+        return Objects.equals(value, nickname.value);
     }
 
     @Override
