+++ src/main/java/woowacourse/shoppingcart/exception/bodyexception/IllegalEmailException.java
@@ -1,4 +1,4 @@
-package woowacourse.shoppingcart.exception;
+package woowacourse.shoppingcart.exception.bodyexception;
 
 public class IllegalEmailException extends IllegalRequestException {
