+++ src/main/java/woowacourse/shoppingcart/exception/bodyexception/IllegalNicknameException.java
@@ -1,4 +1,4 @@
-package woowacourse.shoppingcart.exception;
+package woowacourse.shoppingcart.exception.bodyexception;
 
 public class IllegalNicknameException extends IllegalRequestException {
