+++ src/main/java/woowacourse/shoppingcart/exception/bodyexception/IllegalQuantityException.java
@@ -1,4 +1,4 @@
-package woowacourse.shoppingcart.exception;
+package woowacourse.shoppingcart.exception.bodyexception;
 
 public class IllegalQuantityException extends IllegalRequestException {
