+++ src/main/java/woowacourse/shoppingcart/exception/bodyexception/IllegalRequestException.java
@@ -1,6 +1,7 @@
-package woowacourse.shoppingcart.exception;
+package woowacourse.shoppingcart.exception.bodyexception;
 
 import org.springframework.http.HttpStatus;
+import woowacourse.shoppingcart.exception.bodyexception.BodyToReturnException;
 
 public class IllegalRequestException extends BodyToReturnException {
