+++ src/main/java/woowacourse/shoppingcart/exception/bodyexception/NotExistProductInCartException.java
@@ -1,4 +1,6 @@
-package woowacourse.shoppingcart.exception;
+package woowacourse.shoppingcart.exception.bodyexception;
+
+import woowacourse.shoppingcart.exception.bodyexception.IllegalRequestException;
 
 public class NotExistProductInCartException extends IllegalRequestException {
