+++ src/main/java/woowacourse/shoppingcart/exception/nobodyexception/UnauthorizedTokenException.java
@@ -1,6 +1,7 @@
-package woowacourse.shoppingcart.exception;
+package woowacourse.shoppingcart.exception.nobodyexception;
 
 import org.springframework.http.HttpStatus;
+import woowacourse.shoppingcart.exception.nobodyexception.NotBodyToReturnException;
 
 public class UnauthorizedTokenException extends NotBodyToReturnException {
