+++ src/test/java/woowacourse/shoppingcart/dao/CartItemDaoTest.java
@@ -12,7 +12,7 @@ import org.springframework.boot.test.autoconfigure.jdbc.JdbcTest;
 import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.test.context.TestConstructor;
 import org.springframework.test.context.jdbc.Sql;
-import woowacourse.shoppingcart.domain.Cart;
+import woowacourse.shoppingcart.domain.cart.Cart;
 
 @JdbcTest
 @AutoConfigureTestDatabase(replace = Replace.NONE)
