+++ src/test/java/woowacourse/shoppingcart/domain/EmailTest.java
@@ -5,7 +5,8 @@ import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
-import woowacourse.shoppingcart.exception.IllegalEmailException;
+import woowacourse.shoppingcart.domain.customer.Email;
+import woowacourse.shoppingcart.exception.bodyexception.IllegalEmailException;
 
 public class EmailTest {
