+++ src/test/java/woowacourse/shoppingcart/domain/QuantityTest.java
@@ -5,7 +5,8 @@ import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
-import woowacourse.shoppingcart.exception.IllegalQuantityException;
+import woowacourse.shoppingcart.domain.cart.Quantity;
+import woowacourse.shoppingcart.exception.bodyexception.IllegalQuantityException;
 
 public class QuantityTest {
